backend/requirements.txt
# BSSOD Analyzer Backend API
# Install with: pip install -r requirements.txt

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# HTTP Client for AI API
httpx>=0.27.0

# Environment Variables
python-dotenv>=1.0.0

# File Upload Handling
python-multipart>=0.0.12

# Data Validation
pydantic>=2.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
